Use class regex to find Java files as well
Basically we do this to classify a file as
being part of the test or the regular build:

if filname.endswith('.java') and
   re.match( config[-istest],
             filename.replace('.java','.class') ):
   print("we consider this part of the test")

As you are allowed to modify the regex in
config['-istest'] this should be sufficient
even when you have "special needs", i.e.
assuming you have a class named:

    WhateverTest.java

as part of your normal build, this would get
filtered out by the default config['-istest']
regex, which basically ignores anything that
ends with 'Test.java', but you could replace
it with a regex like this:

    -istest "^((?!Whatever).)*Test\\\.class$"

which would allow your normal test classes
to be picked up, but allow WhateverTest.java
to be part of the normal build.